+++ examples/get-jobs.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-require_once __DIR__ . '/../vendor/autoload.php';
-
 use DjThossi\ErgosoftSdk\Domain\BaseUrl;
 use DjThossi\ErgosoftSdk\ErgosoftFactory;
 use DjThossi\ErgosoftSdk\SimpleErgosoftConfiguration;
+
+require_once __DIR__ . '/../vendor/autoload.php';
 
 // Initialize the Ergosoft API client with YOUR URL and optional timeout (defaults to 10 seconds)
 $configuration = new SimpleErgosoftConfiguration(new BaseUrl('https://YOUR_API_URL'));