+++ tests/Integration/FactoryTest.php
@@ -9,6 +9,7 @@ use DjThossi\ErgosoftSdk\Api\CancelRippingJobApi;
 use DjThossi\ErgosoftSdk\Api\DeleteJobApi;
 use DjThossi\ErgosoftSdk\Api\GetJobByGuidApi;
 use DjThossi\ErgosoftSdk\Api\GetJobsApi;
+use DjThossi\ErgosoftSdk\Api\MoveUpJobPositionInQueueApi;
 use DjThossi\ErgosoftSdk\Api\SubmitDeltaXmlFileApi;
 use DjThossi\ErgosoftSdk\Api\SubscribeJobStatusApi;
 use DjThossi\ErgosoftSdk\Api\TestCommunicationsApi;
@@ -90,6 +91,14 @@ class FactoryTest extends TestCase
 
         /* @noinspection UnnecessaryAssertionInspection */
         $this->assertInstanceOf(TestCommunicationsApi::class, $factory->createTestCommunicationsApi());
+    }
+
+    public function testCreateMoveUpJobPositionInQueueApi(): void
+    {
+        $factory = $this->createErgosoftFactory();
+
+        /* @noinspection UnnecessaryAssertionInspection */
+        $this->assertInstanceOf(MoveUpJobPositionInQueueApi::class, $factory->createMoveUpJobPositionInQueueApi());
     }
 
     private function createErgosoftFactory(): ErgosoftFactory